🔨 Refactoring: Skeleton 컴포넌트 폴더 분리
