+++ src/components/Skeleton/SkeletonFollowUserInfo/index.ts
@@ -1,0 +1,3 @@
+import SkeletonFollowUserInfo from './SkeletonFollowUserInfo';
+
+export { SkeletonFollowUserInfo };
