+++ src/components/Skeleton/SkeletonPostDetail/PostDetailSkeleton.style.ts
