+++ src/components/Skeleton/SkeletonPostDetail/index.ts
@@ -1,0 +1,3 @@
+import PostDetailSkeleton from './PostDetailSkeleton';
+
+export { PostDetailSkeleton };
