+++ src/components/Skeleton/SkeletonPostPreview/index.ts
@@ -1,0 +1,3 @@
+import PostPreviewSkeleton from './PostPreviewSkeleton';
+
+export { PostPreviewSkeleton };
