+++ src/components/Skeleton/SkeletonProfile/index.ts
@@ -1,0 +1,3 @@
+import ProfileSkeleton from './ProfileSkeleton';
+
+export { ProfileSkeleton };
