+++ src/components/Skeleton/SkeletonSearch/SearchFollowSkeleton.style.ts
