+++ src/components/Skeleton/SkeletonSearch/index.ts
@@ -1,0 +1,4 @@
+import SearchFollowSkeleton from './SearchFollowSkeleton';
+import SearchPostPreviewSkeleton from './SearchPostPreviewSkeleton';
+
+export { SearchFollowSkeleton, SearchPostPreviewSkeleton };
